createMetadata mutation and query loader
The idea is that we implement queries in static .graphql files, then
we can read and cache these with a simple utility function.

For this to be possible, we need to handle variables in queries, so
static queries can be written separately and the variables can
be passed at runtime to the query function.