client: extend authentication to support API token
API token is actually a tenant token, so there is no need to specify the
tenant to be used, neither to request a new tenant token as API tokens
have a longer validity period. (Also it is not possible to renew an API
token.)

When accessing the API using a tenant token, we can query the tenant name
using a GraphQL query, this way can test the access and acquire the tenant
name as well.